Visual_Language_Navigation_ManiMob/requirements.txt
numpy
ftfy
regex
tqdm
hydra-core==1.0.5
wandb==0.10.15
matplotlib
torchvision==0.8.2
pytorch-lightning==1.1.0
tqdm
Pillow
networks
requests
pandas
nltk
ipdb
